backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
passlib
python-multipart
docx
PyPDF2
cloudinary